+++ src/fastmath/fields.clj
@@ -1,7 +1,48 @@
 (ns fastmath.fields
-  ""
-  {:metadoc/categories {:comp "Compose fields"
-                        :cr "Create fields"}}
+  "Vector field functions.
+
+  Vector fields are functions R^2->R^2.
+
+  Names are taken from fractal flames world where such fields are call `variations`. Most implementations are taken from [JWildfire](http://jwildfire.org/) software.
+
+  ## Creation
+
+  To create vector field call [[field]] multimethod with name of the field as keyword.
+
+  Some of the vector fields require additional configuration as a map of parameters as keywords and values. Call [[parametrization]] to create random one or to merge with provided.
+
+  Additionally you can provide `amount` parameter which is scaling factor for vector field (default: `1.0`).
+
+  ## Derived fields
+
+  You can use several method to derive new vector field from the other one(s). Possible options are:
+
+  * [[derivative]], [[grad-x]], [[grad-y]] - directional derivative of the field
+  * [[sum]] - sum of two fields
+  * [[multiply]] - multiplication of the fields
+  * [[composition]] - composition of the fields
+  * [[angles]] - angles of the field vectors
+
+  ## Scalar fields
+
+  You can derive scalar fields from given vector field(s):
+
+  * [[jacobian]] - determinant of jacobian matrix
+  * [[divergence]] - divergence of the field
+  * [[cross]] - cross product of the fields (as a determinant of the 2x2 matrix of vectors)
+  * [[dot]] - dot product
+  * [[angle-between]] - angle between vectors from fields.
+
+  ## Combinations
+
+  The other option is to create vector field using some of the above possibilities. Combination is a tree of field operations with parametrizations. Functions:
+
+  * [[combine]] - create vector field randomly of from given parametrization.
+  * [[random-configuration]] - returns random configuration as a map
+  * [[randomize-configuration]] - change parametrization for given configuration."
+  {:metadoc/categories {:cr "Create fields"
+                        :sc "Derive scalar field from vector field"
+                        :vf "Derive vector field from other vector field(s)."}}
   (:require [fastmath.core :as m]
             [fastmath.complex :as c]
             [fastmath.random :refer :all]
@@ -854,7 +895,7 @@
 
 ;; ### CircleRand
 
-(make-config-method circlerand {:Sc (srandom 0.1 1.2)
+(make-config-method circlerand {:Sc (drand 0.1 1.2)
                                 :Dens (drand 1)
                                 :X (drand -20 20)
                                 :Y (drand -20 20)
@@ -876,12 +917,12 @@
                                (v/mult 2.0)
                                (v/applyf #(inc ^double %))
                                (v/mult Sc)))]
-          (if (and (< iter 100)
-                   (or (> (discrete-noise (+ Seed (.x MN)) (.y MN)) Dens)
-                       (> ^double (v/mag XY) (-> (discrete-noise (+ 10.0 (.x MN)) (+ 3.0 (.y MN)))
-                                                 (* 0.7)
-                                                 (+ 0.3)
-                                                 (* Sc)))))
+          (if (bool-and (< iter 60)
+                        (bool-or (> (discrete-noise (+ Seed (.x MN)) (.y MN)) Dens)
+                                 (> ^double (v/mag XY) (-> (discrete-noise (+ 10.0 (.x MN)) (+ 3.0 (.y MN)))
+                                                           (* 0.7)
+                                                           (+ 0.3)
+                                                           (* Sc)))))
             (recur (inc iter))
             (-> MN
                 (v/mult 2.0)
@@ -3074,40 +3115,142 @@
 ;;
 (def ^{:dynamic true
        :doc "When random configuration for [[combine]] is used. Skip vector fields which are random."
-       :metadoc/categories #{:comp}}
+       :metadoc/categories #{:vf}}
   *skip-random-fields* false)
 
-(defn derivative
-  "Calculate derivative of fn using `h` (default: `1.0e-6`) as a step."
-  {:metadoc/categories #{:comp}}
+(defn- directional-derivative
+  "Compute directional derivative."
+  [f dir ^double amount ^double h]
+  (fn [v]
+    (let [v1 (f v)
+          v2 (f (v/add v dir))]
+      (v/mult (v/div (v/sub v2 v1) h) amount))))
+
+(defn derivative 
+  "Calculate directional derivative of fn. Derivative is calculated along [1,1] vector with `h` as a step (default `1.0e-6`)."
+  {:metadoc/categories #{:vf}}
   ([f ^double amount ^double h]
-   (let [^Vec2 d (Vec2. h h)]
-     (fn [^Vec2 v]
-       (let [v1 (f v)
-             v2 (f (v/add v d))]
-         (v/mult (v/div (v/sub v2 v1) h) amount)))))
+   (directional-derivative f (Vec2. h h) amount h))
   ([f ^double h]
    (derivative f 1.0 h))
   ([f]
    (derivative f 1.0 1.0e-6)))
 
+(defn grad-x
+  "Calculate gradient along x axis."
+  {:metadoc/categories #{:vf}}
+  ([f ^double amount ^double h]
+   (directional-derivative f (Vec2. h 0.0) amount h))
+  ([f ^double h]
+   (grad-x f 1.0 h))
+  ([f]
+   (grad-x f 1.0 1.0e-6)))
+
+(defn grad-y
+  "Calculate gradient along y axis."
+  {:metadoc/categories #{:vf}}
+  ([f ^double amount ^double h]
+   (directional-derivative f (Vec2. 0.0 h) amount h))
+  ([f ^double h]
+   (grad-y f 1.0 h))
+  ([f]
+   (grad-y f 1.0 1.0e-6)))
+
+(defn jacobian
+  "Det of Jacobian of the field"
+  {:metadoc/categories #{:sc}}
+  ([f] (jacobian f 1.0e-6))
+  ([f ^double h]
+   (let [gx (grad-x f h)
+         gy (grad-y f h)]
+     (fn [^Vec2 v]
+       (let [^Vec2 gxv (gx v)
+             ^Vec2 gyv (gy v)]
+         (- (* (.x gxv) (.y gyv))
+            (* (.x gyv) (.y gxv))))))))
+
+(defn divergence
+  "Divergence of the field"
+  {:metadoc/categories #{:sc}}
+  ([f] (divergence f 1.0e-6))
+  ([f ^double h]
+   (let [gx (grad-x f h)
+         gy (grad-y f h)]
+     (fn ^double [v]
+       (let [^Vec2 gxv (gx v)
+             ^Vec2 gyv (gy v)]
+         (+ (.x gxv) (.y gyv)))))))
+
+(defn magnitude
+  "Magnitude of the vectors from field."
+  {:metadoc/categories #{:sc}}
+  [f]
+  (fn ^double [v]
+    (v/mag (f v))))
+
+(defn heading
+  "Angle of the vectors from field."
+  {:metadoc/categories #{:sc}}
+  [f]
+  (fn ^double [v]
+    (v/heading (f v))))
+
+(defn- generate-scalar-field
+  "Generate scalar field"
+  [op]
+  (fn
+    ([f]
+     (fn [v]
+       (op (f v) v)))
+    ([f1 f2]
+     (fn [v]
+       (op (f1 v) (f2 v))))))
+
+(def ^{:doc "2d cross product (det of the 2x2 matrix) of the input vector and result of the vector field.
+
+In case when two vector fields are given, cross product is taken from results of vector fields."
+       :metadoc/categories #{:sc}}
+  cross (generate-scalar-field v/cross))
+(def ^{:doc "Dot product of the input vector and result of the vector field.
+
+In case when two vector fields are given, cross product is taken from result of vector fields."
+       :metadoc/categories #{:sc}}
+  dot (generate-scalar-field v/dot))
+(def ^{:doc "Angle between input vector and result of the vector field.
+
+In case when two vector fields are given, cross product is taken from result of vector fields.
+
+Resulting value is from range `[-PI,PI]`."
+       :metadoc/categories #{:vf}}
+  angle-between (generate-scalar-field v/angle-between))
+
+(defn scalar->vector-field
+  "Returns vector field build from scalar fields of the input vector and result of the vector field."
+  {:metadoc/categories #{:vf}}
+  ([scalar f] 
+   (fn [v]
+     (Vec2. (scalar (f v)) (scalar v))))
+  ([scalar f1 f2]
+   (fn [v]
+     (Vec2. (scalar (f1 v)) (scalar (f2 v))))))
+
 (defn composition
   "Compose two vector fields."
-  {:metadoc/categories #{:comp}}
+  {:metadoc/categories #{:vf}}
   ([f1 f2 ^double amount]
    (fn [v] (v/mult (f1 (f2 v)) amount)))
   ([f1 f2] (composition f1 f2 1.0)))
 
 (defn sum
   "Add two vector fields."
-  {:metadoc/categories #{:comp}}
+  {:metadoc/categories #{:vf}}
   ([f1 f2 ^double amount]
    (fn [v] (v/mult (v/add (f1 v) (f2 v)) amount)))
   ([f1 f2] (sum f1 f2 1.0)))
 
 (defn multiplication
   "Multiply two vector fields (as a element-wise multiplication of results)."
-  {:metadoc/categories #{:comp}}
+  {:metadoc/categories #{:vf}}
   ([f1 f2 ^double amount]
    (fn [v] (v/mult (v/emult (f1 v) (f2 v)) amount)))
   ([f1 f2] (multiplication f1 f2 1.0)))
@@ -3121,7 +3264,7 @@
 (defn- build-random-parametrization-step
   "Create parametrization tree"
   ([f1 f2]
-   (let [operand (rand-nth [:comp :add :comp :add :comp :mult :comp :comp :comp])]
+   (let [operand (rand-nth [:comp :add :comp :add :comp :mult :comp :angles :comp])]
      {:type :operation :name operand :var1 f1 :var2 f2}))
   ([f]
    (randval 0.1 f
@@ -3133,19 +3276,19 @@
 
 (defn randomize-configuration
   "Randomize values for given configuration. Keeps structure untouched."
-  {:metadoc/categories #{:comp}}
+  {:metadoc/categories #{:vf}}
   ([f]
    (if (= (:type f) :variation)
      (assoc f :amount 1.0 :config (parametrization (:name f) {}))
      (let [name (:name f)]
        (if (= name :deriv)
          (assoc f :amount 1.0 :step (m/sq (drand 0.01 1.0)) :var (randomize-configuration (:var f)))
-         (let [amount1 (if (= name :comp) 1.0 (drand -2.0 2.0))
-               amount2 (if (= name :comp) 1.0 (drand -2.0 2.0)) 
+         (let [amount1 (if (#{:comp :angles} name) 1.0 (drand -2.0 2.0))
+               amount2 (if (#{:comp :angles} name) 1.0 (drand -2.0 2.0)) 
                amount (case name
                         :add (/ 1.0 (+ (m/abs amount1) (m/abs amount2)))
                         :mult (/ 1.0 (* amount1 amount2))
-                        :comp 1.0)]
+                        1.0)]
            (assoc f :amount amount
                   :var1 (assoc (randomize-configuration (:var1 f)) :amount amount1)
                   :var2 (assoc (randomize-configuration (:var2 f)) :amount amount2))))))))
@@ -3153,8 +3296,10 @@
 (defn random-configuration
   "Create random configuration for [[combine]] function. Optionally with depth (0 = only root is created).
 
-  See [[combine]] for structure."
-  {:metadoc/categories #{:comp}}
+  See [[combine]] for structure.
+
+  Bind `*skip-random-fields*` to true to exclude fields which are random."
+  {:metadoc/categories #{:vf}}
   ([] (random-configuration (lrand 5)))
   ([depth] (random-configuration depth (build-random-variation-step)))
   ([^long depth f]
@@ -3163,7 +3308,9 @@
      f)))
 
 (defn combine
-  "Create composite vector field function based on configuration (default random).
+  "Create composite vector field function based on configuration
+
+  Call without argument to get random vector field.
 
   Configuration is a tree structure where nodes are one of the following
 
@@ -3185,9 +3332,10 @@
   * `:add` - sum of two variations
   * `:mult` - multiplication
   * `:comp` - composition
+  * `:angles` - vector field from angles
 
   See [[random-configuration]] for example."
-  {:metadoc/categories #{:comp}}
+  {:metadoc/categories #{:vf}}
   ([{:keys [type name amount config var step var1 var2]}]
    (if (= type :variation)
      (field name amount config)
@@ -3198,5 +3346,6 @@
          (case name
            :comp (composition v1 v2 amount)
            :add (sum v1 v2 amount)
-           :mult (multiplication v1 v2 amount))))))
+           :mult (multiplication v1 v2 amount)
+           :angles (scalar->vector-field v/heading v1 v2))))))
   ([] (combine (random-configuration))))